salidas

+++ FrigorificosBle.Almacen.Core/Service/IProductoService.cs
@@ -12,7 +12,7 @@ namespace FrigorificosBle.Almacen.Core.Service
     {
         IList<Linea> GetLineas();
         IList<Medida> GetMedidas();
-        Producto GetById(Int32 id);
+        Producto GetById(long id);
         void Save(Producto producto);
         IEnumerable<Producto> Query(ProductoQueryDto dto);
     }
